Budget module refactor - readability (names / spaces / simplify constructions / etc) - proper types (int for block height, CAmount for money) - prettify log output / drop c_str() there - read budgets on dump before writing (safety) - microoptimization (++it) - sort proposals by Yeas - sort finalized budgets by votes - fix few small bugs

+++ src/masternode-payments.h
@@ -26,8 +26,8 @@ extern std::map<uint256, CMasternodeBlockPayees> mapMasternodeBlocks;
 
 void ProcessMessageMasternodePayments(CNode* pfrom, std::string& strCommand, CDataStream& vRecv);
 bool IsReferenceNode(CTxIn& vin);
-bool IsBlockPayeeValid(const CTransaction& txNew, int64_t nBlockHeight);
-std::string GetRequiredPaymentsString(int64_t nBlockHeight);
+bool IsBlockPayeeValid(const CTransaction& txNew, int nBlockHeight);
+std::string GetRequiredPaymentsString(int nBlockHeight);
 bool IsBlockValueValid(const CBlock& block, int64_t nExpectedValue);
 void FillBlockPayee(CMutableTransaction& txNew, int64_t nFees);
